+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Allow super-admins to change objects' owners [#3260](https://github.com/opendatateam/udata/pull/3260)
+- Fix contact point permissions [#3253](https://github.com/opendatateam/udata/pull/3253)
 
 ## 10.0.8 (2025-01-31)