+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow super-admins to change objects' owners [#3260](https://github.com/opendatateam/udata/pull/3260)
 - Fix contact point permissions [#3253](https://github.com/opendatateam/udata/pull/3253)
+- Fix wrong marshal for community resource uploads [#3261](https://github.com/opendatateam/udata/pull/3261)
 
 ## 10.0.8 (2025-01-31)